requirements.txt
argparse
os
sys
json
re
logging
requests
python-dotenv
google-api-python-client
mutagen
supabase
spotipy
